refactor: rename RecordingWindow to ScreenRecordingWindow (#76)
resolves #76
