+++ src/components/feature/ScreenRecordingWindow/RecordingWindow.tsx
