+++ src/components/feature/ScreenRecordingWindow/components/RecordControls.tsx
