+++ src/components/feature/ScreenRecordingWindow/components/RecordData.tsx
