+++ src/components/feature/ScreenRecordingWindow/components/RecordVideo.tsx
