+++ src/components/feature/ScreenRecordingWindow/components/index.ts
