+++ src/components/feature/ScreenRecordingWindow/hooks/index.ts
