+++ src/components/feature/ScreenRecordingWindow/hooks/useRecorderWindow.ts
