+++ src/components/feature/ScreenRecordingWindow/models/recorder.model.ts
