+++ src/components/feature/ScreenRecordingWindow/services/context/recorder.provider.tsx
