+++ src/components/feature/ScreenRecordingWindow/services/context/recorderWindow.provider.tsx
